// src/app/models/contacts.ts
export class Contacts {

    
     id: number;
     salutation: string;
     first_Name: string;
     middle_Name: string;
     last_Name: string;
     mobile_Number: string;
     reports_To: string;
     department: string;
     fax: string;
     pincode: string;
     country: string;
     state: string;
     city: string;
     street: string;
     account_Name: string;
     phone_Number: string;
     email: string;
     address: string;
     title: string;
   
     
}